+++ iOS_SDK/OneSignalSDK/OneSignalInAppMessages/Controller/OSTriggerController.m
@@ -27,7 +27,6 @@
 
 #import "OSTriggerController.h"
 #import "OSInAppMessagingDefines.h"
-#import "OneSignalHelper.h"
 
 @interface OSTriggerController ()
 @property (strong, nonatomic, nonnull) NSMutableDictionary<NSString *, id> *triggers;
