+++ iOS_SDK/OneSignalSDK/OneSignalInAppMessages/Model/OSInAppMessageTag.m
@@ -26,7 +26,6 @@
 */
 
 #import "OSInAppMessageTag.h"
-#import "OneSignalHelper.h"
 
 @implementation OSInAppMessageTag
