+++ iOS_SDK/OneSignalSDK/OneSignalOutcomes/Source/OutcomeEvents/OSSessionManager.h
@@ -54,5 +54,6 @@
 - (void)onNotificationReceived:(NSString * _Nonnull)notificationId;
 - (void)onDirectInfluenceFromNotificationOpen:(AppEntryAction)entryAction withNotificationId:(NSString * _Nonnull)directNotificationId;
 - (void)attemptSessionUpgrade:(AppEntryAction)entryAction;
+- (NSDate *)sessionLaunchTime;
 
 @end
