--- iOS_SDK/OneSignalSDK/Source/OneSignalViewHelper.m
@@ -1,72 +1,0 @@
-/**
- * Modified MIT License
- *
- * Copyright 2016 OneSignal
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * 1. The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * 2. All copies of substantial portions of the Software may only be used in connection
- * with services provided by OneSignal.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-
-#import <Foundation/Foundation.h>
-#import <WebKit/WebKit.h>
-#import "OneSignalViewHelper.h"
-
-@implementation OneSignalViewHelper : NSObject 
-
-/*
- Given a UIDeviceOrientation return the custom enum ViewOrientation
- */
-+ (ViewOrientation)validateOrientation:(UIDeviceOrientation)orientation {
-    if (UIDeviceOrientationIsPortrait(orientation))
-        return OrientationPortrait;
-    else if (UIDeviceOrientationIsLandscape(orientation))
-        return OrientationLandscape;
-    
-    return OrientationInvalid;
-}
-
-/*
- Given a float convert it to the phone scaled CGFloat
- Multiplies float by iPhones scale value
- */
-+ (CGFloat)sizeToScale:(float)size {
-    float scale = [self getScreenScale];
-    return (CGFloat) (size * scale);
-}
-
-/*
- Get currentDevice screen bounds
- Contains point of origin (x, y) and the size (height, width) of the screen
- Changes in regards to the phones current orientation
- */
-+ (CGRect)getScreenBounds {
-    return UIScreen.mainScreen.bounds;
-}
-
-/*
- Get currentDevice screen scale
- Important for specific constraints so that phones of different resolution scale properly
- */
-+ (float)getScreenScale {
-    return UIScreen.mainScreen.scale;
-}
-
-@end
